requirements.txt
appdirs==1.4.4
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
fonttools==4.38.0
gcloud==0.17.0
googleapis-common-protos==1.57.0
httplib2==0.21.0
idna==3.4
Jinja2==3.1.2
jws==0.1.3
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
mne==1.2.3
numpy==1.23.5
oauth2client==3.0.0
packaging==21.3
Pillow==9.3.0
pooch==1.6.0
protobuf==4.21.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pyparsing==3.0.9
Pyrebase==3.0.27
python-dateutil==2.8.2
python-jwt==2.0.1
pytz==2022.6
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.9
scipy==1.9.3
six==1.16.0
sqlparse==0.4.3
tqdm==4.64.1
urllib3==1.26.13
